Improvements to the app layout.
