+++ ui/package.json
@@ -24,6 +24,8 @@
     "ts-protoc-gen": "^0.14.0",
     "typescript": "^4.1.5",
     "web-vitals": "^1.1.0",
+    "xterm": "^4.11.0",
+    "xterm-addon-fit": "^0.5.0",
     "xterm-for-react": "^1.0.4"
   },
   "optionalDependencies": {
