+++ ui/src/App.css
@@ -10,7 +10,7 @@
 
 .contentBackground{
     display: flex;
-    width: 60%;
+    width: 100%;
     height: 100vh;
     /* change the color of the back ground for developing */
     background-color: rgb(255, 255, 255);
