--- ui/src/Terminal.css
@@ -1,4 +1,0 @@
-.Terminal-Container {
-  margin-top: 64px;
-}
-
