+++ ui/src/components/Terminal.css
@@ -1,0 +1,4 @@
+.Terminal {
+    width: 100%;
+    height: 100%;
+}
